add user data to mappings
This patch allows the user to store additional data in the mapping.
This user data is assumed to be independent of the topology and, hence, can
be accessed through non-virtual methods (even when HybridMapping is used).

In order to avoid storing the user data twice, the CachedMapping has been split
into a new NonHybridMapping and the CachedMapping itself. The new user-visible
class is NonHybridMapping. This should also clarify the semantics of the methods
trace and clone, which no longer reside on the CachedMapping but on the
NonHybridMapping.

The impact of this patch on the user is relatively low. You will only notice it
in two situations:
a) Your GeometryTraits class (e.g., GlobalGeometryTraits) is not derived from
   DefaultGeometryTraits. In this case you have to add a type UserData (see
   DefaultGeometryTraits for an example).
b) You are using the CachedMapping directly, i.e., you are not using one of
   the classes GenericGeometry::BasicGeometry, GenericGeometry::Geometry, or
   GenericGeometry::LocalGeomtry. In this case you simply replace CachedMapping
   by NonHybridMapping.

[[Imported from SVN: r55]]