Test using the exported types JacobianTransposed and JacobianInverseTransposed
We cannot assume that these are FieldMatrices anymore.

This patch contains some evil trickery.  In the test we do need FieldMatrices,
e.g., to get operator[] and other niceties.  However I couldn't find a simple
universal way to get a FieldMatrix from JacobianTransposed.  Therefore the
code now multiplies the matrix from the right by the identity.  The result
is a FieldMatrix :-)

[[Imported from SVN: r239]]