A special code paths for the case dim==0
The functionality of the AxisAlignedCubeGeometry class can be
implemented particularly efficient if dim==0 (i.e. the cube is a vertex).
Even though the case already did the right thing in that case,
it was slower than necessary.  For each component the corresponding
entry of the 'axes' bitfield was looked up -- even though they are
all 'false' when dim==0.  Since 'dim' is a compile time constant
it make sense to implement dim==0 as a special case.

Hence, with this patch the AxisAlignedCubeGeometry class can be
used to implement vertices, without any run-time penalty.
This allow to avoid some special casing in grid implementations.

[[Imported from SVN: r353]]