Add composite quadrature rules for simplex elements
This short file implements a general mechanism for constructing
composite quadrature rules.  It takes some quadrature rule
and an integer refinement level as the input.  The reference
simplex is the refined uniformly as prescribed by the desired
level.  The given input quadrature rule is then used on
each subelement.

Currently this only works for simplices.  I'll try to implement
it also for other elements, but there are some fixes in the
virtual refinement code needed first.