Make all constructors of the QuadratureRule class protected
The QuadratureRule class acts as an abstract base class for quadrature
rule implementations (this is even stated in the documentation[sic]).
However, the fact that it is abstract is not enforced anywhere.
Pure virtual methods cannot be used, because the derived class is only
expected to add a constructor.  Making all constructors of QuadratureRule
protected makes sure no object of class QuadratureRule can be created
by itself.

As a consequence, I have do to a few minor changes to the code that
accesses the _quadratureMap map.  std::map::operator[] calls the default
constructor internally, and that is not allowed anymore.

As a matter of fact, this is a bug in the factory: The QuadratureRuleFactory
creates objects derived from QuadratureRule, and hands them to the
QuadratureRules class for storage.  The latter stores them in a map
of QuadratureRule objects, held by value.  Hence the hand-over from
the factory to the container will slice off any additional data that
the derived class may have.  This only does not crash because no
current quadrature rule implementation holds additional data.