Document the sources of the individual simplex quadrature rules
The simplex quadrature rules have originally been taken from the online
archive of quadrature rules at

http://www.cs.kuleuven.ac.be/~nines/research/ecf/ecf.html

However, no more information than that was given in the file.
This patch adds the precise number in the archive for each quadrature
rule.  For each rule it also adds book or research article where
the rule can be found.