[release,automake] Corrects installation directory of refinement documenation.
When the refinement code was move from dune-grid to dune-geometry, this
was never reflected in the installation directories of the documentation.
It still got installed under share/doc/dune-grid. This patch fixes this
and it is finally installed where it is supposed to be, share/doc/dune-geometry.
(cherry picked from commit 59b32983042dc5f80d41e06e66e3c47062aa92c9)

Signed-off-by: Markus Blatt <markus@dr-blatt.de>