jacobianTransposed: handle evaluation at tip of pyramid
Handle the evaluation of the jacobian at the tip of a pyramid. Note
that the derivative there is only defined if the base of the pyramid is
subject to an affine-linear transformation. Non-affine-linear
transformations are not detected and the result is undefined.

(cherry picked from commit 7ade11ecbfbbe327b5ef647f10e2f5c8007b38ab)
Signed-off-by: Christoph Grüninger <gruenich@dune-project.org>