[GenericGeometry::FieldHelper] Find abs() via argument-dependent lookup, or in std.
This is necessary to work with custom number types.  If they follow the
standard they are not allowed to put their overload of abs() into namespace
std.

Also, this avoids issues with the ordering of includes: if std::abs() is used,
the set of candidates for abs() is fixed when the compiler sees the
GenericGeometry::FieldHelper template.  If the header for the custom number
type is included later, any definition of abs() there won't be taken into
account, even if that header puts that definition into namespace std.  This is
not an issue with argumend-dependent lookup.