Fixes conversion of topologyId to unsigned int.
topologyId in MultiLinearGeometry might be an unsigned int or an
integral_constant of an unsigned int. g++-4.4  has problems with
casting integral_constant to the integral value contained.

To fix this we static methods that convert unsigned int and
integral_constant<unsigned int,v> to unsigned int.

This is a working replacement for e7c5d9248 which got reverted by d76daf4