[QuadratureRules] A vector for objects that are not copy-constructible.
This is needed to make QuadratureRules thread-safe.

Note: in g++-4.7 std::deque works for this purpose (although performance was
not tested -- it may better or worse than NoCopyVector).  In g++-4.4
std::deque requires a copy or move constructor on it's elements, which makes
it unsuitable to provide storage for mutexes and call_once_flags.  Versions
4.5 and 4.6 were not tested, neither were clang++ nor icc.