[QuadratureRules] New function QuadratureRules::maxOrder().
This adds the a static function maxOrder(geometryType, quadratureType) to
QuadratureRules.  This information is needed to make QuadratureRules::rule()
thread-safe: we need to know how large to allocate the caching data
structures, so we only need to make the initialization of each data structure
thread-safe, and we can avoid resizing the data structure later.

This needs to be exposed on QuadratureRules since QuadratureRules uses
QuadratureRulesFactory in the constructions of quadrature rules, which in turn
uses TensorProductQuadratureRule, which in turn recursively uses
lower-dimensional QuadratureRules instances.