[bugfix,release] Fix the checkInside method for subentities
Rather surprisingly, the checkInside method for subentities was completely
broken.  One of the template parameters was wrong, and you'd get a compiler
error as soon as you'd use the method for a codim>0 subentity.

The method has been like that at least since 2012, which makes me suspect
that nobody actually needs it.  I don't -- I just stumbled across the bug
by accident.