[MultiLinearGeometry] Do not use cref() as if it is a customization point.
Customization points (like `begin()`) must be found via ADL, which means one
has to do the two-step construct
```
  using std::begin;
  begin(container);
```
Looking at this from a different angle: if a person reading the code sees the
two-step construct, he will likely assume that it is needed to access a
customization point when he is not familiar with the function being invoked.
This leads to confusion if that function is not actually a customization
point, as in this case `cref()`.