[cleanup] use delegating constructors for AffineGeometry
Since we started to require C++11-compatible compilers, we can now make
use of delegating constructors to remove code duplication and make the
intention of the constructors more clear.

AffineGeometry's constructors always come in pairs: one taking a
ReferenceElement and one taking a GeometryType. Since the GeometryType
is only used to look up a ReferenceElement and then proceed identical to
the former constructor, we can instead delegate the actual construction
from the latter to the former constructor.