[bugfix] use std::make_integer_sequence instead of Dune::Std::make_integer_sequence
Rationale: gcc-4.9 coughs up on the specialization of Hybrid::forEach for
std::integer_sequence when fed with the result of Std::make_integer_sequence.
As we require a C++14-compatible compiler (at least to the extent of gcc-4.9),
this patch resolves the issue by using std::make_integer_sequence instead.