[referenceelement] Turn reference element into lightweight object with value semantics
This commit introduces a new type `ReferenceElement` that only stores
a const pointer to an arbitrary implementation (provided as sole
template parameter) and forwards all functionality to that
pointer. The new type itself has value semantics, i.e. it is copyable
and default-constructible. Copying an object of the type is not more
expensive than passing around a reference to the old reference
element.

The interface of the new reference element is mostly identical to the
old one, but it never hands out references (due to the value
semantics). This affects methods returning geometry types and
coordinates.

Note that the new reference element has a different template signature
than the old one, but as it lives in a different namespace, that is
not a problem. The original `namespace Dune` contains a type alias
that does the necessary translation. Moreover, the container type
`ReferenceElements` now also exports the type of the reference
elements as `ReferenceElement`. The container still returns const
references to avoid silent breakage in user code.

The code also adjusts the geometry implementations to work correctly
with the new semantics. This mostly involves the affine geometry,
which is used internally by the reference element implementation, but
also the remaining geometries, which for some reason all offer a free
function `referenceElement(geometry)` that returned a const ref to the
reference element. These functions have been changed to return by
value instead. While this might break user code, I don't think many
users actually called those functions.