[referenceelements] Change the meaning of Dune::ReferenceElement<>
This patch turns `Dune::ReferenceElement<>` into a meta function
equivalent of `referenceElement()`, allowing users to get the type
of reference element for a geometry using
`ReferenceElement<Geometry>`. Internally, this works by simply
dispatching to a `decltype` on a function call to `referenceElement()`
with arguments of the correct types.

Unfortunately, this does not work in general for multiple arguments
yet, because the old template signature requires exactly one type
argument and one `int`. So I've put the fully working version into
`namespace Dune::Transitional` for now, and `Dune::ReferenceElement`
only accepts one type, but still works with the old syntax.

As a bonus, users get a deprecation warning if they keep using the old
syntax.