[GeometryType] Add constexpr static factory functions for GeometryTypes
As we can't make the setup methods `make...()` constexpr at the
moment (GCC 4.9) and because they are inconvenient in constexpr
context anyway, this patch adds corresponding factory functions.

As there is currently no constructor for a singular type of arbitrary
dimension, this patch also adds a new constructor that allows
explicitly specifying the bool flag for a singular type.