[GeometryType] Add namespace GeometryTypes with constexpr variables and factory functions
As suggested by Martin, this patch adds a new namespace
`GeometryTypes` that contains constexpr objects for the most common
geometry types. As we can't have objects for cubes, simplices and
none, there are instead constexpr functions for the purpose.

This brings the interface in line with what we do for the partition
sets in dune-grid.