[bugfix] Restore reference semantics of ReferenceElement methods
By using decltype(auto) we forward references of the
implementation as references. This reverts the interface
change introduced in !52 in case the implementation returns
references. As a consequence the all tests in dune-localfunctions
pass again.

For (possible) implementations returning values these will
also be forwarded. In the future (after a grace period of
one release), this should be switched to references completely
as proposed in !52.