[GeometryType] Add an integral GeometryType::Id type for use as a template parameter
While GeometryType is constexpr and can thus be used in a lot of compile time contexts, C++ does not
allow us to declare a template parameter of type GeometryType, as the standard only allows data
types that directly map to a built-in integral type for this purpose.

On the other hand, GeometryType itself is just a simple 64 bit struct that stores a few integral
values, so we can easily map it to an integral type.

This patch does just that by adding an opaque nested type GeometryType::Id that is implemented as a
scoped enum as well as implicit conversions between GeometryType and GeometryType::Id. By using a
scoped enum, we clearly tag the type as something which is not a normal number, because C++ requires
explicit conversion between scoped enums and other integral types.

The storage layout of the id is chosen identical to that of GeometryType, making conversion between
the two almost free.

When writing a template with a GeometryType parameter, just use the id instead and convert it back
into a GeometryType inside the template:

template<GeometryType::Id gt_>
class Foo
{
  static constexpr GeometryType gt = gt_;
};