[cleanup] Canonicalize SFINAE in referenceElements()
The used version of SFINAE by enable_if has the advantage
of not adding additional template-parameters. However, it
has the disadvantage of making the generated doxygen documentation
hard to read.

This commit changes this to SFINAE/enable_if by defaulted
int template parameter. This additional parameter also
shows up in doxygen but its much easier to identify as
SFINAE dummy because it follows a simple idomatic pattern.
More importantly it does not mess up the documentation
for the other parameters.