use `std::vector` instead of `NoCopyVector`
`NoCopyVector` is a very limited vector type for items missing a copy
constructor.  This was required in the past as `vector`'s template
parameter `T` "must meet the requirements of CopyAssignable and
CopyConstructible", however since C++11 this was weakened to "The
requirements that are imposed on the elements depend on the actual
operations performed on the container" which should allow use with
non-CopyAssignable and non-CopyConstructible types.