Add contains() and size() method to subentity range
Now the index range returned from referenceElement.subEntities(i,codim,c)
is not just iterable but also provides the methods size() and contains().
This is not just for convenience:

It's a common use case of this method (e.g. when determining boundary DOFs),
to check if a faced contains the subentities associated to DOFs. Until know
you always had to loop over the range provided by referenceelement.subEntity()
and compare resulting indices. Now you can use the contains() method
which just looks up the result in a table.

Design decision:
The per-codim look-up is implemented using a std::bitset<K> with
K=3^dim as upper bound for the per-codim subentity count. While
this is pessipistic, it's much cheaper than any dynamic version
for dim<=4. Feel free to add a better upper bound for larger dimensions.