Added a new quadraturetype.
`GaussJacobi_n_0` is a quadrature type that respects the weight function resulting from conical product.
The rules use already existing quadrature rules compatible with various types
and if the desired order or dimension is too high new rules are computed on demand.
This case requires `LAPACK` and a floating point number type.

The rules currently have an arbitrary "maximal" order of 127. This upper limit is imposed by technical issues.