Guard against misuse of AxisAlignedCubeGeometry constructor
The AxisAlignedCubeGeometry has always had a constructor that was always
intended to be used with dim==coorddim only.  However, this precondition
was never checked anywhere, which could lead to strange bugs.
This patch introduces a static_assert that fails at compile time
if the dim==coorddim precondition is not met.

BUG: https://gitlab.dune-project.org/core/dune-geometry/-/issues/25