Add Impl::toGeometryTypeConstant helper
This helper maps a dynamic `GeometryType` to a compile time
`std::integral_constant<GeometryType::Id,...>`. Thus is functionally
replaces the `IfGeometryType` helper. In contrast to the latter it
does not rely on the undocumented implementation of `GeometryType::Id`
but on `LocalGeometryTypeIndex` and the generic `Hybrid::switchCase`
helper. Furthermore the new helper uses a lambda-friendly callback
interface.